requirements.txt
absl-py==0.6.1
astor==0.7.1
bleach==1.5.0
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
html5lib==0.9999999
idna==2.8
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
numpy==1.15.4
protobuf==3.6.1
pyparsing==2.3.0
python-datamuse==1.2.1
python-dateutil==2.7.5
PyYAML==3.13
requests==2.21.0
scikit-learn==0.20.2
scipy==1.1.0
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
tqdm==4.28.1
urllib3==1.24.1
Werkzeug==0.14.1
wincertstore==0.2
